Implement auto-completions
Make a start with the implementation of completions in both values
and object properties. Often, but not always, we get the correct
'suggestions'. Still need to figure out how to reliably determine
whether or not we need to complete a value or an object property.
Another issue is that the document might be invalid after typing a ':',
which means there won't be a `KsonValue` to analyze. I now solved that
by adding "" to the document. It feels that there should be a more
reliable way to resolve this.