Improve completion handling
Finally, I feel quite good about the way we approach the issue of
reliably offering proper completions and schema info. The 'trick' was to
take into account the tokenized document and find the last typed token.
With this information we know if we need a 'value' completion (so the
key needs to be added to the path), if we are past a token (so for
example the last token is a STRING_CLOSE_QUOTE, then we know that we
should not navigate to that string, but to the 'parent' and get
'property' completions.
This also solves the issue with being on a location 'outside' of the
document, for example on the last empty line of a document. We can now
use/replace the position with the position of the last token.