Improve tests
Improve the tests by using a `<caret>` marker for locations. This makes
the tests a bit more readable and easier to write.
`KsonTextDocumentService.test.ts` has a test that needs to be finished.
We can finish it once we can 'set' teh schema in VS Code per document.
Currently this is hardcoded, so the test would 'assume' a specific
schema.