Update Rust Bindings
The rust bindings were not fully reflecting the API in `Kson.kt`. The
`KsonValue.KsonNumber` in `Kson.kt` is a sealed class with both an
`Decimal` and `Integer` class. In the rust bindings this was
hierarchical.

The rust bindings have been updated to make this hierarchical, at the
cost of some more obscure code (thanks claude). Since we will soon
create the bindings with a bindings generator this extra code isn't too
big of a problem I feel.