Add `forDefinition` to `PathBuilder`
Depending on the type of information we want the path needs to include
the parent. For example if we have the following document:
```kson
top_level_key:
  string_key: "string"
  another_key: "another"
```

If the caret is after the string_key colon then we should include
string_key in the path, both for completions, schema information and
jumping to definitions. if the caret is after the last colon, we want to
provide completions options for all keys. So the path needs to navigate
to the `top_level_key`. For the definitions we want to see the 'string'
definition, and hence don't remove the last element of the path