Add integration test for schema tooling
Features like jumping to definitions, hover information and completions
were hard to test before because we couldn't easily associate schema's
with files. I postponed it until I felt we could test all these features
simultaneously.
This test associates a schema through the status bar, by 'stubbing' the
actions a user would take to associate a schema.
From there we test the different features.
Most of this code has been generated with the following prompt and
Claude:
```
We can associate schema's through the StatusBar in the VSCODE Plugin
(only desktop). I'd like to make an integration test where we create a
workspace, add a schema through the status bar and then test if the
status bar shows the associated schema. We also need to test for
completion, hover, and jumping to definition.
That are features that were already implemented, but couldn't easily be
tested before.
Could you help me implement these tests.
```