Simplify token handling in Lexer and AST processing
This refactor removes the too-early (and sometime clumsy) evaluation
of what the "value" for a token's raw text must be.  We now instead
pass the `sourceTokens` into the AstNode class to whom the parse said
they belong and have those classes own the transformation into a
"value"

This refactor not only streamlines code and organizes transformations
into our Ast class where it definitely seems they belong, it also
positions us to hopefully implement "source maps" for KSON values which
will allow tracing an error on a portion of a KsonValue's value all
the way back to its location in the raw source text (accounting for
things like processed escapes and trimmed embed indents)

One final benefit: this refactor highlights that our embedTag and
embedMeta are improperly hijacking the QuotedString class. I have a
planned improvement that will address, making embedTag and embedMeta
fully compatible with JSON Strings (and hence accommodating fully
idempotent round-trips to the JSON object notation version of an embed)
which will fix this. In the meantime, it's cool to have this refactor
extra-validated by how it helps make obvious subtle problems in the
code.