Refactor KsonValues to wrap AstNodes
KsonValue is the public/friendly interface on a valid KSON AST tree.
The AST API if you will. Having these object hold a pointer to their
source AstNode ensures they have the context they need to provide a
rich interface.  The short term motivation for this refactor is
the upcoming "subLocation" feature we've been laying the groundwork for
in e13178bd85f and 5ceaa3f0b294f0